Fix displaying horizontal scrollbar
By setting width and height to 100%, the pdf-viewer now only takes the horizontal
and vertical space given to it (indirectly through a container element)
by the developer.
Now the horizontal scrollbar isn't at the end of the pdf file but rather at the
end of the pdf-viewer. This changes the amount of space that the viewer will
occupy, thus potentially breaking previous integrations.