Fix "stick to page" bug
When "stickToPage" is true, changing the page didn't result in the page being
actually scrolled into view. This was due to the offsetParent of the page
being the HTMLBodyElement, since it was the only one without "position: static".
By setting the ng2.pdf-viewer-container to "position: relative" it is recognized
as offsetParent of the page and thus the page gets scrolled into view.