feat(pdfjs): upgrade to v4
This is a continuation of PR #1105 which seems by all accounts to have been abandoned by the original author. However, the notable difference is that instead of pushing out a breaking change & forcing an upgrade to Angular 17 simply for Promise.withResolvers, instead I opted to Pollyfill it directly, as that is the only reason the upgrade was necessary (from what I could tell).

I implemented the suggestions that @shamoon suggested in the original PR comments as well.