build: Upgrade pdfjs-dist to latest version
The following error is thrown when using the library
with the old version of pdfjs-dist:
10:22:02 AM [vite] warning:
.../src/frontend/.angular/cache/18.2.2/vite/deps/ng2-pdf-viewer.js
11940|        const worker = yield import(
11941|          /*webpackIgnore: true*/
11942|          this.workerSrc
   |          ^^^^^^^^^^^^^^
11943|        );
11944|        return worker.WorkerMessageHandler;
The above dynamic import cannot be analyzed by Vite.
See https://github.com/rollup/plugins/tree/master/packages/dynamic-import-vars\#limitations for supported dynamic import formats. If this is intended to be left as-is, you can use the /* @vite-ignore */ comment inside the import() call to suppress this warning.